Revert "feat(ui): sos main leaderboard + crew modal ui (#1206)"
This reverts commit d365ddd10cd0598587ae59460d66d6adc772dfb8.
