+++ src/containers/main/ShellOfSecrets/ShellOfSecrets.tsx
@@ -1,15 +1,13 @@
 import SoSWidget from './SoSWidget/SoSWidget';
 import { useShellOfSecretsStore } from '../../../store/useShellOfSecretsStore';
 import { AnimatePresence } from 'framer-motion';
-import SoSMainModal from './SoSMainModal/SoSMainModal';
 
 export default function ShellOfSecrets() {
-    const { showWidget, showMainModal } = useShellOfSecretsStore();
+    const { showWidget } = useShellOfSecretsStore();
 
     return (
         <>
             <AnimatePresence>{showWidget && <SoSWidget />}</AnimatePresence>
-            <AnimatePresence>{showMainModal && <SoSMainModal />}</AnimatePresence>
         </>
     );
 }
