--- src/containers/main/ShellOfSecrets/SoSMainModal/icons/CloseIcon.tsx
@@ -1,22 +1,0 @@
-const CloseIcon = () => (
-    <svg width="14" height="14" viewBox="0 0 14 14" fill="none" xmlns="http://www.w3.org/2000/svg">
-        <rect
-            x="0.939453"
-            y="2.35786"
-            width="2.0077"
-            height="16.4631"
-            transform="rotate(-45 0.939453 2.35786)"
-            fill="currentColor"
-        />
-        <rect
-            x="2.35938"
-            y="13.9982"
-            width="2.0077"
-            height="16.4631"
-            transform="rotate(-135 2.35938 13.9982)"
-            fill="currentColor"
-        />
-    </svg>
-);
-
-export default CloseIcon;
