--- src/containers/main/ShellOfSecrets/SoSMainModal/sections/SoSMainContent/CrewMining/CrewList/data.ts
@@ -1,166 +1,0 @@
-const data = {
-    members: [
-        {
-            image: 'https://robohash.org/user1',
-            isNew: true,
-            isOnline: true,
-            miningRate: 12.5,
-            id: 'm001',
-            name: 'Alice Cooper',
-        },
-        {
-            image: 'https://robohash.org/user2',
-            isNew: true,
-            isOnline: true,
-            miningRate: 8.7,
-            id: 'm002',
-            name: 'Bob Wilson',
-        },
-        {
-            image: 'https://robohash.org/user3',
-            isNew: false,
-            isOnline: true,
-            miningRate: 15.2,
-            id: 'm003',
-            name: 'Carol Smith',
-        },
-        {
-            image: 'https://robohash.org/user4',
-            isNew: false,
-            isOnline: true,
-            miningRate: 9.8,
-            id: 'm004',
-            name: 'David Brown',
-        },
-        {
-            image: 'https://robohash.org/user5',
-            isNew: false,
-            isOnline: true,
-            miningRate: 11.3,
-            id: 'm005',
-            name: 'Eva Green',
-        },
-        {
-            image: 'https://robohash.org/user6',
-            isNew: false,
-            isOnline: true,
-            miningRate: 7.6,
-            id: 'm006',
-            name: 'Frank Miller',
-        },
-        {
-            image: 'https://robohash.org/user7',
-            isNew: false,
-            isOnline: false,
-            miningRate: 6.4,
-            id: 'm007',
-            name: 'Grace Lee',
-        },
-        {
-            image: 'https://robohash.org/user8',
-            isNew: false,
-            isOnline: false,
-            miningRate: 10.1,
-            id: 'm008',
-            name: 'Henry Ford',
-        },
-        {
-            image: 'https://robohash.org/user9',
-            isNew: false,
-            isOnline: false,
-            miningRate: 13.7,
-            id: 'm009',
-            name: 'Iris West',
-        },
-        {
-            image: 'https://robohash.org/user10',
-            isNew: false,
-            isOnline: false,
-            miningRate: 8.9,
-            id: 'm010',
-            name: 'Jack Ryan',
-        },
-        {
-            image: 'https://robohash.org/user11',
-            isNew: false,
-            isOnline: false,
-            miningRate: 14.2,
-            id: 'm011',
-            name: 'Kate Bishop',
-        },
-        {
-            image: 'https://robohash.org/user12',
-            isNew: false,
-            isOnline: false,
-            miningRate: 9.3,
-            id: 'm012',
-            name: 'Liam Neeson',
-        },
-        {
-            image: 'https://robohash.org/user13',
-            isNew: false,
-            isOnline: false,
-            miningRate: 11.8,
-            id: 'm013',
-            name: 'Maria Hill',
-        },
-        {
-            image: 'https://robohash.org/user14',
-            isNew: false,
-            isOnline: false,
-            miningRate: 7.2,
-            id: 'm014',
-            name: 'Nick Fury',
-        },
-        {
-            image: 'https://robohash.org/user15',
-            isNew: false,
-            isOnline: false,
-            miningRate: 12.9,
-            id: 'm015',
-            name: 'Olivia Pope',
-        },
-        {
-            image: 'https://robohash.org/user16',
-            isNew: false,
-            isOnline: false,
-            miningRate: 8.4,
-            id: 'm016',
-            name: 'Peter Parker',
-        },
-        {
-            image: 'https://robohash.org/user17',
-            isNew: false,
-            isOnline: false,
-            miningRate: 15.6,
-            id: 'm017',
-            name: 'Quinn King',
-        },
-        {
-            image: 'https://robohash.org/user18',
-            isNew: false,
-            isOnline: false,
-            miningRate: 10.5,
-            id: 'm018',
-            name: 'Ron Weasley',
-        },
-        {
-            image: 'https://robohash.org/user19',
-            isNew: false,
-            isOnline: false,
-            miningRate: 13.1,
-            id: 'm019',
-            name: 'Sara Lance',
-        },
-        {
-            image: 'https://robohash.org/user20',
-            isNew: false,
-            isOnline: false,
-            miningRate: 9.7,
-            id: 'm020',
-            name: 'Tom Hardy',
-        },
-    ],
-};
-
-export default data;
