--- src/containers/main/ShellOfSecrets/SoSMainModal/sections/SoSMainContent/CrewMining/icons/GreenArrowIcon.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-const GreenArrowIcon: React.FC = () => (
-    <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 20 20" fill="none">
-        <path
-            d="M7.21822 4.37128C7.47995 4.01598 7.98022 3.94006 8.3356 4.20178C8.61163 4.40509 8.87402 4.60839 9.10258 4.78676C9.55874 5.14279 10.1698 5.63291 10.783 6.16546C11.3924 6.6945 12.0224 7.28116 12.5063 7.82874C12.7474 8.10165 12.9691 8.38375 13.1349 8.65882C13.2875 8.91192 13.4557 9.25971 13.4557 9.64018C13.4557 10.0207 13.2875 10.3685 13.1349 10.6216C12.9691 10.8967 12.7474 11.1787 12.5063 11.4516C12.0224 11.9993 11.3924 12.5859 10.783 13.115C10.1698 13.6475 9.55874 14.1376 9.10258 14.4937C8.87402 14.672 8.61164 14.8753 8.33561 15.0786C7.98022 15.3404 7.47995 15.2645 7.21822 14.9091C7.11301 14.7662 7.06236 14.6 7.0625 14.4352L7.0625 9.64018L7.0625 4.84526C7.06236 4.68048 7.11301 4.51417 7.21822 4.37128Z"
-            fill="#C5FF47"
-        />
-    </svg>
-);
-
-export default GreenArrowIcon;
