--- src/containers/main/ShellOfSecrets/SoSMainModal/sections/SoSMainContent/CrewMining/icons/PlusIcon.tsx
@@ -1,15 +1,0 @@
-const PlusIcon = () => (
-    <svg width="93" height="93" viewBox="0 0 93 93" fill="none" xmlns="http://www.w3.org/2000/svg">
-        <rect x="45.418" y="35.1641" width="2.73142" height="24.0365" fill="white" />
-        <rect
-            x="34.7656"
-            y="48.5469"
-            width="2.73142"
-            height="24.0365"
-            transform="rotate(-90 34.7656 48.5469)"
-            fill="white"
-        />
-    </svg>
-);
-
-export default PlusIcon;
