--- src/containers/main/ShellOfSecrets/SoSMainModal/sections/SoSMainContent/Leaderboard/LeaderboardList/data.ts
@@ -1,164 +1,0 @@
-export const data = {
-    entries: [
-        {
-            handle: '@X_Miner_Pro',
-            status: 'mining',
-            last_mined: null,
-            duration: '52d 21h 32m 15s',
-            rank: 1,
-            image: 'https://robohash.org/user1',
-        },
-        {
-            handle: '@CryptoKing',
-            status: 'idle',
-            last_mined: '2024-12-08T10:08:00Z',
-            duration: '52d 21h 32m 15s',
-            rank: 2,
-            image: 'https://robohash.org/user2',
-        },
-        {
-            handle: '@BlockMaster',
-            status: 'mining',
-            last_mined: null,
-            duration: '52d 21h 32m 15s',
-            rank: 3,
-            image: 'https://robohash.org/user3',
-        },
-        {
-            handle: '@HashRateHero',
-            status: 'idle',
-            last_mined: '2024-12-08T10:08:00Z',
-            duration: '52d 21h 32m 15s',
-            rank: 4,
-            image: 'https://robohash.org/user4',
-        },
-        {
-            handle: '@MiningPro',
-            status: 'mining',
-            last_mined: null,
-            duration: '52d 21h 32m 15s',
-            rank: 5,
-            image: 'https://robohash.org/user5',
-        },
-        {
-            handle: '@CoinCollector',
-            status: 'mining',
-            last_mined: null,
-            duration: '52d 21h 32m 15s',
-            rank: 6,
-            image: 'https://robohash.org/user6',
-        },
-        {
-            handle: '@BlockchainBoss',
-            status: 'idle',
-            last_mined: '2024-12-08T10:08:00Z',
-            duration: '52d 21h 32m 15s',
-            rank: 7,
-            image: 'https://robohash.org/user7',
-        },
-        {
-            handle: '@HashMaster',
-            status: 'mining',
-            last_mined: null,
-            duration: '52d 21h 32m 15s',
-            rank: 8,
-            image: 'https://robohash.org/user8',
-        },
-        {
-            handle: '@CryptoChamp',
-            status: 'idle',
-            last_mined: '2024-12-08T10:08:00Z',
-            duration: '52d 21h 32m 15s',
-            rank: 9,
-            image: 'https://robohash.org/user9',
-        },
-        {
-            handle: '@MineKing',
-            status: 'mining',
-            last_mined: null,
-            duration: '52d 21h 32m 15s',
-            rank: 10,
-            image: 'https://robohash.org/user10',
-        },
-        {
-            handle: '@BlockWarrior',
-            status: 'mining',
-            last_mined: null,
-            duration: '52d 21h 32m 15s',
-            rank: 11,
-            image: 'https://robohash.org/user11',
-        },
-        {
-            handle: '@HashHunter',
-            status: 'idle',
-            last_mined: '2024-12-08T10:08:00Z',
-            duration: '52d 21h 32m 15s',
-            rank: 12,
-            image: 'https://robohash.org/user12',
-        },
-        {
-            handle: '@CryptoNinja',
-            status: 'mining',
-            last_mined: null,
-            duration: '52d 21h 32m 15s',
-            rank: 13,
-            image: 'https://robohash.org/user13',
-        },
-        {
-            handle: '@BlockExplorer',
-            status: 'idle',
-            last_mined: '2024-12-08T10:08:00Z',
-            duration: '52d 21h 32m 15s',
-            rank: 14,
-            image: 'https://robohash.org/user14',
-        },
-        {
-            handle: '@MiningMaster',
-            status: 'mining',
-            last_mined: null,
-            duration: '52d 21h 32m 15s',
-            rank: 15,
-            image: 'https://robohash.org/user15',
-        },
-        {
-            handle: '@CryptoElite',
-            status: 'mining',
-            last_mined: null,
-            duration: '52d 21h 32m 15s',
-            rank: 16,
-            image: 'https://robohash.org/user16',
-        },
-        {
-            handle: '@BlockGenius',
-            status: 'idle',
-            last_mined: '2024-12-08T10:08:00Z',
-            duration: '52d 21h 32m 15s',
-            rank: 17,
-            image: 'https://robohash.org/user17',
-        },
-        {
-            handle: '@HashLegend',
-            status: 'mining',
-            last_mined: null,
-            duration: '52d 21h 32m 15s',
-            rank: 18,
-            image: 'https://robohash.org/user18',
-        },
-        {
-            handle: '@CryptoWizard',
-            status: 'idle',
-            last_mined: '2024-12-08T10:08:00Z',
-            duration: '52d 21h 32m 15s',
-            rank: 19,
-            image: 'https://robohash.org/user19',
-        },
-        {
-            handle: '@MiningLegend',
-            status: 'mining',
-            last_mined: null,
-            duration: '52d 21h 32m 15s',
-            rank: 20,
-            image: 'https://robohash.org/user20',
-        },
-    ],
-};
