--- src/containers/main/ShellOfSecrets/SoSMainModal/sections/SoSMainContent/PrizeDrops/PrizeDrops.tsx
@@ -1,12 +1,0 @@
-import MacbookDrop from './Drops/MacbookDrop';
-import UpcomingDrop from './Drops/UpcomingDrop';
-import { Wrapper } from './styles';
-
-export default function PrizeDrops() {
-    return (
-        <Wrapper>
-            <MacbookDrop />
-            <UpcomingDrop />
-        </Wrapper>
-    );
-}
