--- src/containers/main/ShellOfSecrets/SoSMainModal/sections/SoSMainContent/styles.ts
@@ -1,12 +1,0 @@
-import styled from 'styled-components';
-
-export const Wrapper = styled('div')`
-    width: 100%;
-    height: 100%;
-
-    display: flex;
-    flex-direction: column;
-    gap: 20px;
-
-    padding-top: 20px;
-`;
