--- src/containers/main/ShellOfSecrets/SoSMainModal/sections/SoSMainSidebar/SoSMainSidebar.tsx
@@ -1,14 +1,0 @@
-import MainTimer from './MainTimer/MainTimer';
-import SuperCharger from './SuperCharger/SuperCharger';
-import { ContentLayer, Wrapper } from './styles';
-
-export default function SoSMainSidebar() {
-    return (
-        <Wrapper>
-            <ContentLayer>
-                <MainTimer />
-                <SuperCharger />
-            </ContentLayer>
-        </Wrapper>
-    );
-}
