--- src/containers/main/ShellOfSecrets/SoSMainModal/sections/SoSMainSidebar/SuperCharger/icons/KeyIcon.tsx
@@ -1,39 +1,0 @@
-const KeyIcon = () => (
-    <svg xmlns="http://www.w3.org/2000/svg" width="89" height="89" viewBox="0 0 89 89" fill="none">
-        <rect
-            opacity="0.5"
-            x="0.874767"
-            y="44.8242"
-            width="61.2401"
-            height="61.2401"
-            transform="rotate(-45 0.874767 44.8242)"
-            stroke="white"
-            strokeOpacity="0.5"
-            strokeWidth="1.177"
-        />
-        <rect
-            x="10.2231"
-            y="44.8242"
-            width="48.0215"
-            height="48.0215"
-            transform="rotate(-45 10.2231 44.8242)"
-            stroke="white"
-            strokeOpacity="0.5"
-            strokeWidth="2.354"
-        />
-        <rect
-            x="15.4102"
-            y="44.8242"
-            width="40.6879"
-            height="40.6879"
-            transform="rotate(-45 15.4102 44.8242)"
-            fill="#2E4224"
-        />
-        <path
-            d="M37.517 33.4492C41.8015 33.4492 45.2843 36.9333 45.2843 41.2158C45.2843 41.9893 45.1201 42.7493 44.8894 43.4884L53.2958 51.8926V51.8938C53.3871 51.9851 53.4389 52.1073 53.4414 52.2368V56.6513C53.4414 56.924 53.2193 57.1448 52.9478 57.1448L48.5372 57.1374C48.4076 57.1374 48.2829 57.0881 48.1904 56.9968L46.4873 55.293C46.396 55.2004 46.3454 55.0758 46.3454 54.9463V53.3634H44.7609C44.6276 53.3646 44.4993 53.3115 44.4055 53.2166L43.3701 52.1814C43.2763 52.0877 43.2245 51.9606 43.2257 51.8286V50.2444H41.6424C41.5128 50.2444 41.3882 50.1939 41.2944 50.1038L39.7764 48.5962C39.0409 48.8207 38.2857 48.9823 37.5144 48.9823C33.2298 48.9823 29.7458 45.4983 29.7458 41.2158C29.7458 39.074 30.6158 37.1332 32.0227 35.728C33.4295 34.3215 35.3721 33.4493 37.5144 33.4493L37.517 33.4492ZM35.672 37.401C35.1475 37.401 34.6453 37.6095 34.275 37.9797C33.9048 38.3498 33.6963 38.8519 33.6963 39.375C33.6963 39.8982 33.9048 40.4003 34.275 40.7704C34.6453 41.1406 35.1475 41.3491 35.672 41.3491C36.7617 41.3491 37.6453 40.4657 37.6453 39.375C37.6453 38.2844 36.7617 37.401 35.672 37.401Z"
-            fill="#D4E5CC"
-        />
-    </svg>
-);
-
-export default KeyIcon;
