+++ src/containers/main/ShellOfSecrets/components/Scanlines/Scanlines.tsx
@@ -1,31 +1,18 @@
 import { useEffect, useRef } from 'react';
 import { Canvas, Vignette } from './styles';
 
-interface ScanlinesProps {
-    scaleToWindow?: boolean;
-}
-
-const Scanlines: React.FC<ScanlinesProps> = ({ scaleToWindow = false }) => {
+const Scanlines = () => {
     const canvasRef = useRef<HTMLCanvasElement>(null);
 
     useEffect(() => {
         const canvas = canvasRef.current;
         if (!canvas) return;
 
-        if (scaleToWindow) {
-            canvas.width = window.innerWidth / 3;
-            canvas.height = window.innerHeight / 3;
-        }
-
         const ctx = canvas.getContext('2d');
         if (!ctx) return;
 
         let animationFrameId: number;
-        let scanlineOffset = 0.2; // Adjusted initial scanlineOffset
-
-        // Create ImageData once and reuse it
-        const imageData = ctx.createImageData(canvas.width, canvas.height);
-        const data = imageData.data;
+        let scanlineOffset = 0;
 
         const drawScanlines = () => {
             if (!ctx || !canvas) return;
@@ -33,22 +20,24 @@ const Scanlines: React.FC<ScanlinesProps> = ({ scaleToWindow = false }) => {
             ctx.clearRect(0, 0, canvas.width, canvas.height);
 
             // Draw TV static
+            const imageData = ctx.createImageData(canvas.width, canvas.height);
+            const data = imageData.data;
             for (let i = 0; i < data.length; i += 4) {
-                const noise = Math.random() < 0.5 ? 19 : 28; // Updated noise colors
+                const noise = Math.random() * 255;
                 data[i] = noise;
                 data[i + 1] = noise;
                 data[i + 2] = noise;
-                data[i + 3] = 255;
+                data[i + 3] = 255 * 0.08;
             }
             ctx.putImageData(imageData, 0, 0);
 
             // Draw scanlines
             for (let y = scanlineOffset; y < canvas.height; y += 3) {
-                ctx.fillStyle = '#1c212b'; // Lighter scanline color
+                ctx.fillStyle = 'rgba(255, 255, 255, 0.08)';
                 ctx.fillRect(0, y, canvas.width, 1);
             }
 
-            scanlineOffset = (scanlineOffset + 0.1) % 3;
+            scanlineOffset = (scanlineOffset + 0.1) % 4;
 
             animationFrameId = requestAnimationFrame(drawScanlines);
         };
@@ -58,11 +47,11 @@ const Scanlines: React.FC<ScanlinesProps> = ({ scaleToWindow = false }) => {
         return () => {
             cancelAnimationFrame(animationFrameId);
         };
-    }, [scaleToWindow]);
+    }, []);
 
     return (
         <>
-            <Canvas ref={canvasRef} style={{ mixBlendMode: 'screen', backgroundColor: '#131519' }} />
+            <Canvas ref={canvasRef} style={{ mixBlendMode: 'screen' }} />
             <Vignette />
         </>
     );
