+++ src/containers/main/ShellOfSecrets/components/Scanlines/styles.ts
@@ -4,7 +4,7 @@ export const Canvas = styled('canvas')`
     position: absolute;
     inset: 0;
     pointer-events: none;
-    z-index: 1;
+    z-index: 2;
     width: 100%;
     height: 100%;
 `;
@@ -13,7 +13,7 @@ export const Vignette = styled.div`
     position: absolute;
     inset: 0;
     pointer-events: none;
-    z-index: 2;
-    background: radial-gradient(circle, rgba(0, 0, 0, 0.3) 0%, rgba(18, 18, 19, 0.8) 100%);
+    z-index: 1;
+    background: radial-gradient(circle at center, rgba(0, 0, 0, 0) 0%, rgba(0, 0, 0, 0.5) 100%);
     mix-blend-mode: multiply;
 `;
