+++ src/store/useShellOfSecretsStore.ts
@@ -1,6 +1,5 @@
 import { CrewMember } from '@app/types/ws.ts';
 import { create } from './create.ts';
-import { AwardedTimeBonus } from '@app/types/sosTypes.ts';
 
 const SOS_GAME_ENDING_DATE = new Date('2025-01-30');
 export const MINING_EVENT_INTERVAL_MS = 15000;
@@ -34,7 +33,6 @@ interface State {
     showWidget: boolean;
     totalBonusTimeMs: number;
     revealDate: Date;
-    showMainModal: boolean;
     wsConnectionState: WsConnectionState;
 }
 
@@ -42,10 +40,8 @@ interface Actions {
     setReferrals: (referrals: ReferralsResponse) => void;
     setShowWidget: (showWidget: boolean) => void;
     setTotalBonusTimeMs: (totalTimeBonusUpdate: number) => void;
-    setTotalTimeBonus: (totalTimeBonusUpdate: AwardedTimeBonus) => void;
-    getTimeRemaining: () => { days: number; hours: number; totalRemainingMs: number; minutes: number; seconds: number };
-    setShowMainModal: (showMainModal: boolean) => void;
     registerWsConnectionEvent: (event: WsConnectionEvent) => void;
+    getTimeRemaining: () => { days: number; hours: number; totalRemainingMs: number };
 }
 
 const initialState: State = {
@@ -53,7 +49,6 @@ const initialState: State = {
     showWidget: false,
     totalBonusTimeMs: 0,
     revealDate: SOS_GAME_ENDING_DATE,
-    showMainModal: false,
     wsConnectionState: {
         state: 'off',
     },
@@ -64,24 +59,6 @@ export const useShellOfSecretsStore = create<State & Actions>()((set, get) => ({
     setReferrals: (referrals) => set({ referrals }),
     setShowWidget: (showWidget) => set({ showWidget }),
     setTotalBonusTimeMs: (totalTimeBonusUpdate: number) => set({ totalBonusTimeMs: totalTimeBonusUpdate }),
-    setTotalTimeBonus: (totalTimeBonusUpdate: AwardedTimeBonus) => {
-        let newTotalTimeBonus = 0;
-        if (totalTimeBonusUpdate.days) {
-            newTotalTimeBonus += totalTimeBonusUpdate.days * 24 * 60 * 60 * 1000;
-        }
-        if (totalTimeBonusUpdate.hours) {
-            newTotalTimeBonus += totalTimeBonusUpdate.hours * 60 * 60 * 1000;
-        }
-        if (totalTimeBonusUpdate.minutes) {
-            newTotalTimeBonus += totalTimeBonusUpdate.minutes * 60 * 1000;
-        }
-        if (totalTimeBonusUpdate.seconds) {
-            newTotalTimeBonus += totalTimeBonusUpdate.seconds * 1000;
-        }
-        if (newTotalTimeBonus !== get().totalBonusTimeMs) {
-            set({ totalBonusTimeMs: newTotalTimeBonus });
-        }
-    },
     registerWsConnectionEvent: (event: WsConnectionEvent) =>
         set(({ wsConnectionState }) => {
             if (event.state === 'off' || event.state === 'up') {
@@ -114,9 +91,6 @@ export const useShellOfSecretsStore = create<State & Actions>()((set, get) => ({
 
         const days = Math.floor(remainingMs / (1000 * 60 * 60 * 24));
         const hours = Math.floor((remainingMs % (1000 * 60 * 60 * 24)) / (1000 * 60 * 60));
-        const minutes = Math.floor((remainingMs % (1000 * 60 * 60)) / (1000 * 60));
-        const seconds = Math.floor((remainingMs % (1000 * 60)) / 1000);
-        return { days, hours, totalRemainingMs: remainingMs, minutes, seconds };
+        return { days, hours, totalRemainingMs: remainingMs };
     },
-    setShowMainModal: (showMainModal) => set({ showMainModal }),
 }));
